Return a list instead of bare path string.
GitHub-Issue: 10
https://github.com/emacs-php/phpstan.el/issues/10

This is a simple mistake. This function requires a string list, this
pattern assigns just a string to the variable.
Since a character string is a type of sequence, the append function
treats it as a vector of char(int) instead of type error.

This issue was reported by @kwvanderlinde, thanks!